requirements.txt
python-dotenv
langchain_groq
langchain
langchain_core
fastapi
uvicorn